Full-frontend/requirements.txt
Flask==3.0.0
faiss-cpu==1.7.4
Jinja2==3.1.2
google-auth==2.21.0
google-auth-oauthlib==1.0.0
numpy==1.26.0
pandas==2.0.3
Pillow==10.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
reportlab==3.6.13
scikit-learn==1.3.1
sentence-transformers==2.2.2
transformers==4.29.2
huggingface-hub==0.20.3
spacy[en_core_web_sm]==3.7.4
torch==2.0.1
urllib3==1.26.18
google-generativeai==0.8.5
python-dotenv==1.1.1
requests==2.32.4
xhtml2pdf==0.2.11
